scripts/neo4j/requirements.txt
# Neo4j Knowledge Graph Dependencies
neo4j>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
supabase>=2.0.0
